Add productData.js
